Move events to better directories
