+++ src/Impostor.Api/Events/Player/PlayerChatEvent.cs
@@ -1,18 +1,21 @@
 ﻿using Impostor.Api.Games;
 using Impostor.Api.Net.Inner.Objects;
 
-namespace Impostor.Api.Events.Net
+namespace Impostor.Api.Events.Player
 {
-    public class PlayerDestroyedEvent : IGameEvent
+    public class PlayerChatEvent : IGameEvent
     {
-        public PlayerDestroyedEvent(IGame game, IInnerPlayerControl playerControl)
+        public PlayerChatEvent(IGame game, IInnerPlayerControl playerControl, string message)
         {
             Game = game;
             PlayerControl = playerControl;
+            Message = message;
         }
 
         public IGame Game { get; }
 
         public IInnerPlayerControl PlayerControl { get; }
+
+        public string Message { get; }
     }
 }
