+++ src/Impostor.Api/Events/Player/PlayerSpawnedEvent.cs
@@ -1,21 +1,18 @@
 ﻿using Impostor.Api.Games;
 using Impostor.Api.Net.Inner.Objects;
 
-namespace Impostor.Api.Events
+namespace Impostor.Api.Events.Player
 {
-    public class PlayerChatEvent : IGameEvent
+    public class PlayerSpawnedEvent : IGameEvent
     {
-        public PlayerChatEvent(IGame game, IInnerPlayerControl playerControl, string message)
+        public PlayerSpawnedEvent(IGame game, IInnerPlayerControl playerControl)
         {
             Game = game;
             PlayerControl = playerControl;
-            Message = message;
         }
 
         public IGame Game { get; }
 
         public IInnerPlayerControl PlayerControl { get; }
-
-        public string Message { get; }
     }
 }
