+++ src/Impostor.Plugins.Debugger/Pages/Index.razor
@@ -3,6 +3,7 @@
 @using Impostor.Api.Events.Managers
 @using Impostor.Api.Games.Managers
 @using Impostor.Api.Events
+@using Impostor.Api.Events.Player
 @implements Impostor.Api.Events.IEventListener
 @inject IEventManager EventManager
 @inject IGameManager GameManager
