+++ src/Impostor.Plugins.Example/Handlers/GameEventListener.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Impostor.Api.Events;
+using Impostor.Api.Events.Player;
 
 namespace Impostor.Plugins.Example.Handlers
 {
