+++ src/Impostor.Plugins.Example/Impostor.Plugins.Example.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
   </PropertyGroup>
-
+  
   <ItemGroup>
     <ProjectReference Include="..\Impostor.Api\Impostor.Api.csproj"/>
   </ItemGroup>
