+++ src/Impostor.Server/Net/Inner/Objects/InnerPlayerControl.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Impostor.Api;
 using Impostor.Api.Events;
 using Impostor.Api.Events.Managers;
-using Impostor.Api.Events.Net;
+using Impostor.Api.Events.Player;
 using Impostor.Api.Innersloth;
 using Impostor.Api.Net;
 using Impostor.Api.Net.Inner.Objects;
