+++ src/Impostor.Server/Net/State/Game.State.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using Impostor.Api;
 using Impostor.Api.Events;
+using Impostor.Api.Events.Player;
 using Impostor.Api.Innersloth;
 using Impostor.Api.Net;
 using Impostor.Api.Net.Messages;
